requirements.txt
beautifulsoup4==4.12.3
python_dateutil==2.8.2
Requests==2.31.0
